Implement initial advanced text rendering logic
There are currently two view modes:

 * highlights: all bold tags are displayed. If no bold tags exist, show up to
   100 characters.

 * unrestricted: show all text.

 * Also thought of introducing a `condensed' view whereby only _one_ bold tag
   would be shown, in addition to `n' words each way. This view is presently
   disabled.

A 'more/less' button is also shown aligned to the right margin which can be
toggled between the 'highlights' and 'unrestricted' modes.

The text returned in `sentences' seems to have passed through a buggy
filter/processing function as often words are repeated and bolding of keywords
is applied incorrectly.