Ensure only one request of each kind if active at any time
Logic has been put in place to make sure that only one request is active at any
one time. This means that the API methods `moreTexts' and `getBinData' allow
each only one request simultaneously. Given that when a request is already
ongoing and a (promise) contract necessarily already exists, `null' is returned
instead when an additional request is initiated as a clear indication that the
method refuses to engage in a further contract until the current one
finishes. I'm not entirely sure we should still return a promise and then reject
it in this particular instance.