Don't use `dragenter'
Implementation is buggy on both Chrome and Firefox in that often times (MUCH more so
on Firefox) a `dragleave' event is triggered even though the pointer is clearly
still inside the drop area BUT, crucially, the `dragenter' isn't triggered
again. This results in the drop area losing its special drop style whilst the
pointer is still within its bounds.