Cater for IE's special needs
Yet another IE bug (or feature, according to the Redmond morons) that took me
awhile to identify and (sort-of) fix: the contents of the data set in the
`dragstart' event via `setData' can ONLY be queried in the `drop' event. This,
of course, creates a number of problems, as illustrated in the code. This commit
is a mere temporary workaround. A proper fix will be worked on tomorrow.