Notify client only when component has fully reset
Sorting Desk now has the ability to track ongoing async requests and, as a
result, can notify its client when it has fully reset its state. This is of
little to no practical use in production but allows us to craft highly reliable
tests and in a much easier fashion too.