Add document rating example page.
This makes a query to a Diffeo server endpoint to find profiles, then
brings up a list of documents that may be about those entities.  The name
of the entity is also used as a keyword for a Wikipedia search.  For each
article, the user rates it based on whether or not it is about the entity.

This example also replaces the SortingDesk CSS with Bootstrap, and shows
how SortingDesk might be integrated into a Knockout application.