HACK: Use a space in the event transfer data.
For whatever reason, the data in the `Text` event transfer data is being
inserted into `contenteditable` divs. (It wasn't before, although I
suppose it makes sense that it is.)

So we now transmit the id of the thing being dragged/dropped via
`DossierId` so that it doesn't get thrown into the text. BUT, we still
need to set `Text` to something so that the browser believes that it's
something that's really droppable. An empty string won't due, but a
single space will. :-)

Only tested in Chrome 39.