LabelBrowser: require `callbacks´ and reference to API instance
The LabelBrowser component no longer requires a reference to the running
`SortingDesk´ instance and can therefore be executed irrespective of whether
Sorting Desk is running or not.

In addition, the `LabelBrowser´ component now requires a `callbacks´ object map,
similar to the one used by `SortingQueue´ but its implementation is rather crude
and very much adhoc. It, and other components such as `BinExplorer´, sorely need
a callbacks handler class.

The `Ui´ class also received some improvements to the code, mainly to do with
structure. Functionally, nothing changed.