Chrome: create resource loader mechanism
There is a limitation to loading certain resource types, such as custom fonts,
that is imposed by the fact that the extension's own id is not known until it is
installed.

At present, this resource loader class supports injection of custom fonts
into the active tab's DOM, but can be extended to other resource types.