SortingCommon: remove `dbg´ class
The idea behind this class was so we could switch between production and
development modes and, depending on the active mode, control which messages are
logged. It could even be further improved so as to show more information in a
custom format.

However, there was a problem with the way I had implemented the class, which
essentially meant that crucial attributes such as the line number and file name
of the caller were replaced by the callee's. This could be solved with some
simple trickery involving "stack" manipulation but would take some time to
implement. For now, and given that I do not think this is a priority, I have
reverted to using the traditional browser `console´ object.