Do not remove items with fancy dismissal.
I think one of the nice advantages of "fancy dismissal" is that it gives
the caller a bit more control as to how an item gets dismissed. I think
this should probably include whether the item even gets dismissed at
all, or rather, *when* it gets dismissed.

Use case: consider that a label is added in response to a dismissal.
Once the item is removed from the queue, a new search is started
immediately. If `SortingQueue` dismisses the item before the backend
adds the label, the search might not have the label, which means it
might show the result that the user just dismissed. (The backend hides
items from the search results based on the presence of a label.)
If `SortingQueue` *doesn't* dismiss the item, then the caller can wait
to dismiss the item until a label has been confirmed to be added by the
backend, which means the next search is guaranteed to factor that label
into filtering.

This breaks other uses of fancy dismissal that relied on `SortingQueue`
doing the dismissal automatically. These clients should add this code to
their handler for the `dismissed` event:

    item.owner.remove(item);

[breaking-change]