chrome: use automatically injected content scripts
This is a regression, as we were using manually injected content scripts
*before* pages loaded.  Manually injecting scripts is no longer viable
due to a Chrome bug that prevents the background script from receiving
tab update notifications in *some* instances and there isn't anything
that can be done to fix or work around it.

This commit also eliminates use of page action and re-introduces the
concept of "browser action" because the gawd-like uber-dudes at Google
know best than anyone else and prevented simultaneous use of browser and
page actions.  Having a browser action makes it clear to the user that
they can interact with it, whereas the page action wasn't so clear. The
browser action should allow the user to easily activate/deactivate
Sorting Desk by clicking on the extension's button.