sorting-desk: fix potential loading status bug
This bug could occur if loading the subfolder list happens faster than
it takes to increment the loading status (previously it was set inside a
`setTimeout´ call).

Also fixed by this commit is clearing of the loading status regardless
after the request completes (it was being cleared on success only).