extensions: add Sorting Queue customisations
This commit adds a source file containing Sorting Queue customisations
that browser extensions depend on.  The only customisation available at
the moment is for search result items.

Support was also added for scores containing 5 star icons instead of the
raw score number.  A non-empty star is shown per 1/5 of the rounded-up
score.  The raw score number is still displayed as a tooltip.