sorting-queue: improve workflow of search results retrieval
Before this change, Sorting Queue required the `moreTexts´ callback to
return a convenient array containing the search results.

This commit changes that and, instead, Sorting Queue requires a map to
be returned from the `moreTexts´ callback *containing* a `results´ array
property.  It then triggers an event just before rendering and
passes *all* the data -- not just `results´ -- received from the
`moreTexts´ callback to allow clients to perform any processing before
rendering begins.  Once rendering is over, the data is dropped.

To fix any existing systems, simply adjust the `moreTexts´ callback such
that instead of returning an array of search results, it instead returns
a map with a `results´ array property in it, e.g.:

Before: return [ { item0 }, { item 1 } ... ];
Now:    return {
          results: [ { item0 }, { item 1 } ... ]
        };

[breaking-change]