Use different event when redrawing.
I think it's useful to distinguish between a real update---which requires
asking the client for more items---and a redraw based on a filter.

If you previously had code that assumes `items-updated` was fired on a
redraw, you'll need to change that code to bind the same function to
the `items-redrawn` event. `items-updated` is still fired otherwise.

[breaking-change]