sorting-desk: item status
Explorer entities -- a folder, subfolder or text/image item -- can now
be displayed a status on a per row basis.  See `sd.openquery.Controller`
for an illustration of this.

In addition, the OpenQuery feature now makes use of this feature and
sets the operation status on the affected subfolder when a query is
initiated and when it terminates.